docs(VTreeview): change item-disabled example to use item-props
fixes #19882
