+++ packages/docs/src/examples/v-treeview/prop-item-props.vue
@@ -1,0 +1,86 @@
+<template>
+  <v-treeview
+    :items="items"
+    item-props
+    selectable
+  ></v-treeview>
+</template>
+
+<script setup>
+  import { ref } from 'vue'
+
+  const items = ref([
+    {
+      id: 1,
+      title: 'Applications :',
+      disabled: true,
+      children: [
+        { id: 2, title: 'Calendar : app' },
+        { id: 3, title: 'Chrome : app' },
+        { id: 4, title: 'Webstorm : app' },
+      ],
+    },
+    {
+      id: 5,
+      title: 'Documents :',
+      children: [
+        {
+          id: 6,
+          title: 'vuetify :',
+          children: [
+            {
+              id: 7,
+              title: 'src :',
+              disabled: true,
+              children: [
+                { id: 8, title: 'index : ts' },
+                { id: 9, title: 'bootstrap : ts' },
+              ],
+            },
+          ],
+        },
+        {
+          id: 10,
+          title: 'material2 :',
+          children: [
+            {
+              id: 11,
+              title: 'src :',
+              children: [
+                { id: 12, title: 'v-btn : ts' },
+                { id: 13, title: 'v-card : ts' },
+                { id: 14, title: 'v-window : ts' },
+              ],
+            },
+          ],
+        },
+      ],
+    },
+    {
+      id: 15,
+      title: 'Downloads :',
+      children: [
+        { id: 16, title: 'October : pdf', disabled: true },
+        { id: 17, title: 'November : pdf', disabled: true },
+        { id: 18, title: 'Tutorial : html', disabled: true },
+      ],
+    },
+    {
+      id: 19,
+      title: 'Videos :',
+      children: [
+        {
+          id: 20,
+          title: 'Tutorials :',
+          children: [
+            { id: 21, title: 'Basic layouts : mp4' },
+            { id: 22, title: 'Advanced techniques : mp4' },
+            { id: 23, title: 'All about app : dir' },
+          ],
+        },
+        { id: 24, title: 'Intro : mov' },
+        { id: 25, title: 'Conference introduction : avi' },
+      ],
+    },
+  ])
+</script>
